chore: use selector for phone-auth-enabled
